HARMONY-3380 [drlvm][winx64][jit] JET operates incorrectly with long values. The regression test will follow shortly.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@518134 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/jet/arith_rt.cpp
@@ -182,7 +182,7 @@ jlong __stdcall rt_h_i64_shift(jlong v1, int v2, JavaByteCodes op)
         return v1 >> (v2&0x3F);
         //  case OPCODE_IUSHR:  return ( v1 >= 0 ) ? v1 >> (v2&0x3F) : (v1 >> (v2&0x3F)) + (2L << ~(v2&0x3F));
     case OPCODE_IUSHR:
-        return ((unsigned long long)v1) >> (v2&0x3F);
+        return ((uint64)v1) >> (v2&0x3F);
     default:
         break;
     }
